pyproject.toml
[project]
name = "nba-ev"
version = "0.1.0"
description = "NBA statistics analysis and visualization package"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "nbformat>=5.7.0",
    "openpyxl>=3.1.0", # For Excel file support
    "httpx>=0.24.0", # For API requests
    "python-dotenv>=1.0.0", # For environment variables
    "basketball-reference-web-scraper>=4.12.1", # For Basketball Reference data
    "nba-api>=1.3.1", # For NBA API data
    "pytest>=8.3.4",
    "selenium>=4.18.1",  # For web scraping
    "webdriver-manager>=4.0.1",  # For managing webdriver installations
    "streamlit>=1.32.0",  # For interactive dashboard
    "plotly>=5.19.0",  # For interactive plots
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pylint>=2.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nba_ev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
]

[tool.pylint.format]
max-line-length = 88
